Organize tests and utils from '@/tests' folder
