+++ eisbuk-admin/src/__tests__/customerDataTriggers.test.ts
@@ -1,5 +1,3 @@
-import firebase from "firebase";
-
 import {
   Category,
   Collection,
@@ -7,10 +5,10 @@ import {
   OrgSubCollection,
 } from "eisbuk-shared";
 
-import { adminDb } from "@/tests/settings";
+import { adminDb } from "@/__testSettings__";
 import { ORGANIZATION } from "@/config/envInfo";
 
-import { deleteAll } from "@/tests/utils";
+import { deleteAll } from "@/__testUtils__/firestore";
 import {
   waitForCondition,
   stripIdAndSecretKey,
@@ -20,18 +18,16 @@ import { testWithEmulator } from "@/__testUtils__/envUtils";
 
 import { saul } from "@/__testData__/customers";
 
-type DocumentData = firebase.firestore.DocumentData;
-
-beforeEach(async () => {
-  await deleteAll();
-});
-
 // string path to `customers` collection
 const customersPath = `${Collection.Organizations}/${ORGANIZATION}/${OrgSubCollection.Customers}`;
 // string path to `bookings` collection
 const bookingsPath = `${Collection.Organizations}/${ORGANIZATION}/${OrgSubCollection.Bookings}`;
 
 describe("Customer triggers", () => {
+  beforeEach(async () => {
+    await deleteAll();
+  });
+
   testWithEmulator(
     "should apply secretKey and customerId (to customer structure) when a customer record is added",
     async () => {
@@ -77,7 +73,7 @@ describe("Customer triggers", () => {
       // wait for `secretKey` to be generated by the data trigger
       const { secretKey } = (await waitForCondition({
         documentPath: `${customersPath}/${saulId}`,
-        condition: hasSecretKey,
+        condition: (data) => data && data.secretKey,
       })) as Customer;
       // wait for customer entry to be added to `bookings` collection
       const bookingsEntry = await waitForCondition({
@@ -103,7 +99,7 @@ describe("Customer triggers", () => {
       // wait for `secretKey` to be created
       const { secretKey } = (await waitForCondition({
         documentPath: `${customersPath}/${saul.id}`,
-        condition: hasSecretKey,
+        condition: (data) => data && data.secretKey,
       })) as Customer;
       // wait for first round of `bookings` entry updates
       await waitForCondition({
@@ -122,14 +118,3 @@ describe("Customer triggers", () => {
     }
   );
 });
-
-/**
- * Check for secretKey in provided document.
- * Used to test `addMissingSecretKey` function for
- * customer registration
- * @param data document data
- * @returns whether secret key exists
- */
-const hasSecretKey = (data: DocumentData | undefined) => {
-  return data && data.secretKey;
-};
