+++ eisbuk-admin/src/components/debugPage.tsx
@@ -5,13 +5,15 @@ import Container from "@material-ui/core/Container";
 import Box from "@material-ui/core/Box";
 import Button from "@material-ui/core/Button";
 
+import { ORGANIZATION } from "@/config/envInfo";
+
+import { __functionsZone__ } from "@/lib/constants";
+
 import { CloudFunction } from "@/enums/functions";
 
 import AppbarAdmin from "@/components/layout/AppbarAdmin";
 
 import useTitle from "@/hooks/useTitle";
-
-import { functionsZone, ORGANIZATION } from "@/config/envInfo";
 
 /**
  * Invokes cloud function
@@ -22,7 +24,7 @@ export const invokeFunction = (functionName: CloudFunction) => {
   return async (): Promise<void> => {
     const res = await firebase
       .app()
-      .functions(functionsZone)
+      .functions(__functionsZone__)
       .httpsCallable(functionName)({ organization: ORGANIZATION });
 
     console.log(res.data);
