+++ eisbuk-admin/src/store/actions/__tests__/bookingOperations.test.ts
@@ -22,7 +22,6 @@ import {
 import { testWithEmulator } from "@/__testUtils__/envUtils";
 import * as firestoreUtils from "@/__testUtils__/firestore";
 import { setupTestBookings } from "../__testUtils__/firestore";
-import { deleteAll } from "@/tests/utils";
 
 const bookingsRef = firestoreUtils
   .getFirebase()
@@ -73,7 +72,7 @@ jest.mock("i18next", () => ({
 describe("Booking Notifications", () => {
   beforeEach(async () => {
     jest.clearAllMocks();
-    await deleteAll();
+    await firestoreUtils.deleteAll();
   });
 
   describe("'bookInterval'", () => {
