--- eisbuk-admin/src/tests/cloudFunctions.test.tsx
@@ -1,79 +1,0 @@
-import { deleteAllCollections, loginWithPhone } from "./utils";
-import { functionsZone } from "@/config/envInfo";
-import { adminDb } from "./settings";
-import firebase from "firebase/app";
-import "firebase/functions";
-
-beforeAll(async () => {
-  await deleteAllCollections(adminDb, ["organizations"]);
-  await firebase.auth().signOut();
-});
-
-const maybeDescribe = process.env.FIRESTORE_EMULATOR_HOST
-  ? describe
-  : xdescribe;
-
-maybeDescribe("Cloud functions", () => {
-  it("can be pinged", async (done) => {
-    const result = await firebase
-      .app()
-      .functions(functionsZone)
-      .httpsCallable("ping")({
-      foo: "bar",
-    });
-    expect(result).toEqual({ data: { pong: true, data: { foo: "bar" } } });
-    done();
-  });
-
-  it("denies access to users not belonging to the organization", async (done) => {
-    await adminDb
-      .collection("organizations")
-      .doc("default")
-      .set({
-        admins: ["test@example.com", "+1234567890"],
-      });
-    // We're not logged in yet, so this should throw
-    await expect(
-      firebase.app().functions(functionsZone).httpsCallable("createTestData")({
-        organization: "default",
-      })
-    ).rejects.toThrow();
-
-    // We log in with the wrong user
-    await loginWithUser("wrong@example.com");
-    await expect(
-      firebase.app().functions(functionsZone).httpsCallable("createTestData")({
-        organization: "default",
-      })
-    ).rejects.toThrow();
-
-    // ...and with the right one
-    await firebase.auth().signOut();
-    await loginWithUser("test@example.com");
-    await firebase
-      .app()
-      .functions(functionsZone)
-      .httpsCallable("createTestData")({
-      organization: "default",
-    });
-
-    // or using the phone number
-    await firebase.auth().signOut();
-    await loginWithPhone("+1234567890");
-    await firebase
-      .app()
-      .functions(functionsZone)
-      .httpsCallable("createTestData")({
-      organization: "default",
-    });
-    done();
-  });
-});
-
-const loginWithUser = async (email: string): Promise<void> => {
-  try {
-    await firebase.auth().createUserWithEmailAndPassword(email, "secret");
-  } catch (e) {
-    await firebase.auth().signInWithEmailAndPassword(email, "secret");
-  }
-};
